+++ live.go
@@ -19,9 +19,9 @@ import (
 	"time"
 
 	"github.com/dgraph-io/dgo/v240/protos/api"
-	"github.com/dgraph-io/dgraph/v24/chunker"
-	"github.com/dgraph-io/dgraph/v24/filestore"
-	"github.com/dgraph-io/dgraph/v24/x"
+	"github.com/hypermodeinc/dgraph/v24/chunker"
+	"github.com/hypermodeinc/dgraph/v24/filestore"
+	"github.com/hypermodeinc/dgraph/v24/x"
 	"github.com/pkg/errors"
 	"golang.org/x/sync/errgroup"
 )
@@ -246,7 +246,7 @@ func (l *liveLoader) uid(ns uint64, val string) (string, error) {
 		return uid, nil
 	}
 
-	asUID, err := l.n.db.LeaseUIDs(1)
+	asUID, err := l.n.engine.LeaseUIDs(1)
 	if err != nil {
 		return "", fmt.Errorf("error allocating UID: %w", err)
 	}