+++ CHANGELOG.md
@@ -10,6 +10,11 @@
 - fix: alter schema on reverse edge after querying schema
   [#55](https://github.com/hypermodeinc/modusDB/pull/55)
 
+- feat: update interface to engine and namespace
+  [#57](https://github.com/hypermodeinc/modusDB/pull/57)
+
+- chore: Update dgraph dependency [#62](https://github.com/hypermodeinc/modusDB/pull/62)
+
 ## 2025-01-02 - Version 0.1.0
 
 Baseline for the changelog.